feat: audioMeta의 chunkId필드를 Long으로 변경

+++ src/main/java/com/jolupbisang/demo/application/meeting/dto/AudioMeta.java
@@ -6,7 +6,7 @@ public record AudioMeta(
         String type,
         long userId,
         long meetingId,
-        int chunkId,
+        long chunkId,
         LocalDateTime timestamp,
         String encoding
 ) {
